requirements.txt
anyascii==0.3.1
asgiref==3.4.1
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cloudinary==1.26.0
cssselect2==0.4.1
Django==3.2.6
django-admin-interface==0.17.1
django-admin-sortable2==1.0
django-appconf==1.0.4
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-colorfield==0.4.3
django-cors-headers==3.8.0
django-dbbackup==3.3.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-extensions==3.1.3
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-material==1.9.0
django-merged-inlines==1.0.3
django-model-utils==4.2.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-reverse-admin==2.9.6
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fonttools==4.28.5
html5lib==1.1
idna==3.2
l18n==2021.3
mysqlclient==2.0.3
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.7
pdf2image==1.16.0
Pillow==8.3.1
pyactiveresource==2.2.2
pycairo==1.20.1
pycparser==2.21
pydotplus==2.0.2
pydyf==0.1.2
PyJWT==2.1.0
pyparsing==3.0.3
pyphen==0.12.0
pytz==2021.1
PyVimeo==0.3.10
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
ShopifyAPI==10.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.2.1
telepath==0.2
tinycss2==1.1.1
urllib3==1.26.6
wagtail==3.0.1
wagtailfontawesome==1.2.1
weasyprint==54.0
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
xxhash==0.4.3
zopfli==0.1.9
